+++ client/src/pages/admin/createProduct.jsx
@@ -1,252 +1,252 @@
 import axios from 'axios'
 import { useState } from 'react'
-
-
-const validate=(form)=> {
-    let error = {};
-    
-    if (!form.name) {
-        error.name = "enter name";
-    }
-    
-   
-    if (!form.manufacturer) {
-        error.manufacturer = "enter info";
-    }
-    if (!form.origin) {
-        error.origin = "enter info"
-    }
-    if (!form.price) {
-        error.price = "enter info"
-    }
-    if (!form.discount) {
-        error.discount = "enter info"
-    }
-    if (!form.stock) {
-        error.stock = "enter info"
-    }
-    return error;
-  }
   
 
 
 
-export default function Form() {
+export default function CreateProduct() {
+
+
+    const validate=(form)=> {
+        let error = {};
+        
+        if (!form.name) {
+            error.name = "enter name";
+        }
+        
+       
+        if (!form.manufacturer) {
+            error.manufacturer = "enter info";
+        }
+        if (!form.origin) {
+            error.origin = "enter info"
+        }
+        if (!form.price) {
+            error.price = "enter info"
+        }
+        if (!form.discount) {
+            error.discount = "enter info"
+        }
+        if (!form.stock) {
+            error.stock = "enter info"
+        }
+        return error;
+    }
  
-  const categories = [
-  {
-      name: 'Hardware',
-      isMainCategory: true,
-      
-      subcategories: [
-          {
-              name: 'Procesadores (CPU)',
-          },
-          {
-              name: 'Placas Madre (Motherboards)',
-          },
-          {
-              name: 'Placas de Video (GPU)',
-          },
-          {
-              name: 'Almacenamiento',
-              subcategories: [
-                  {
-                      name: 'Disco Rígido Externo',
-                  },
-                  {
-                      name: 'Disco Interno Mecánico',
-                  },
-                  {
-                      name: 'Disco SSD / SSD M.2',
-                  },
-                  {
-                      name: 'Ópticos (DVD / Blu-Ray)',
-                  },
-                  {
-                      name: 'Pen Drives',
-                  },
-                  {
-                      name: 'MicroSD',
-                  },
-              ],
-          },
-          {
-              name: 'Cooling / Refrigeración',
-              subcategories: [
-                  {
-                      name: 'Coolers CPU',
-                  },
-                  {
-                      name: 'Coolers Gabinete',
-                  },
-                  {
-                      name: 'Pastas Térmicas',
-                  },
-              ],
-          },
-          {
-              name: 'Memorias RAM',
-              subcategories: [
-                  {
-                      name: 'DIMM (DDR3, DDR4)',
-                  },
-                  {
-                      name: 'SO-DIMM',
-                  },
-              ],
-          },
-          {
-              name: 'Gabinetes, Fuentes y Alimentación',
-              subcategories: [
-                  {
-                      name: 'Gabinetes',
-                  },
-                  {
-                      name: 'Fuentes',
-                  },
-                  {
-                      name: 'Alimentación',
-                  },
-              ],
-          },
-      ],
-  },
-  {
-      name: 'Monitores',
-      isMainCategory: true,
-      subcategories: [
-          {
-              name: 'Monitores',
-          },
-          {
-              name: 'Monitores Gamer',
-          },
-      ],
-  },
-  {
-      name: 'Periféricos y Accesorios',
-      isMainCategory: true,
-      subcategories: [
-          {
-              name: 'Auriculares',
-          },
-          {
-              name: 'Kit Teclado/Mouse',
-          },
-          {
-              name: 'Teclado',
-          },
-          {
-              name: 'Mouse',
-          },
-          {
-              name: 'Mouse Pads',
-          },
-          {
-              name: 'Micrófonos',
-          },
-          {
-              name: 'Parlantes',
-          },
-          {
-              name: 'Joysticks',
-          },
-          {
-              name: 'Webcams',
-          },
-          {
-              name: 'Conectividad',
-          },
-          {
-              name: 'Tabletas Digitalizadoras',
-          },
-          {
-              name: 'Sillas Gamer',
-          },
-      ],
-  },
-  {
-      name: 'Notebooks / Tablets',
-      isMainCategory: true,
-  },
-  {
-      name: 'Ofertas',
-      isMainCategory: true,
-  },
-];
-const [selectedCategory, setSelectedCategory] = useState('');
-const [selectedSubcategory, setSelectedSubcategory] = useState('');
-  
- 
-  const [form, setForm] = useState({
-    name:"",
-    description:"",
-    manufacturer: "",
-    origin: "",
-    price:"",
-    discount:0,
-    stock: 0,
-    isActive:false,//fata input
-    category:"",
-    image:"",
+    const categories = [
+    {
+        name: 'Hardware',
+        isMainCategory: true,
+        
+        subcategories: [
+            {
+                name: 'Procesadores (CPU)',
+            },
+            {
+                name: 'Placas Madre (Motherboards)',
+            },
+            {
+                name: 'Placas de Video (GPU)',
+            },
+            {
+                name: 'Almacenamiento',
+                subcategories: [
+                    {
+                        name: 'Disco Rígido Externo',
+                    },
+                    {
+                        name: 'Disco Interno Mecánico',
+                    },
+                    {
+                        name: 'Disco SSD / SSD M.2',
+                    },
+                    {
+                        name: 'Ópticos (DVD / Blu-Ray)',
+                    },
+                    {
+                        name: 'Pen Drives',
+                    },
+                    {
+                        name: 'MicroSD',
+                    },
+                ],
+            },
+            {
+                name: 'Cooling / Refrigeración',
+                subcategories: [
+                    {
+                        name: 'Coolers CPU',
+                    },
+                    {
+                        name: 'Coolers Gabinete',
+                    },
+                    {
+                        name: 'Pastas Térmicas',
+                    },
+                ],
+            },
+            {
+                name: 'Memorias RAM',
+                subcategories: [
+                    {
+                        name: 'DIMM (DDR3, DDR4)',
+                    },
+                    {
+                        name: 'SO-DIMM',
+                    },
+                ],
+            },
+            {
+                name: 'Gabinetes, Fuentes y Alimentación',
+                subcategories: [
+                    {
+                        name: 'Gabinetes',
+                    },
+                    {
+                        name: 'Fuentes',
+                    },
+                    {
+                        name: 'Alimentación',
+                    },
+                ],
+            },
+        ],
+    },
+    {
+        name: 'Monitores',
+        isMainCategory: true,
+        subcategories: [
+            {
+                name: 'Monitores',
+            },
+            {
+                name: 'Monitores Gamer',
+            },
+        ],
+    },
+    {
+        name: 'Periféricos y Accesorios',
+        isMainCategory: true,
+        subcategories: [
+            {
+                name: 'Auriculares',
+            },
+            {
+                name: 'Kit Teclado/Mouse',
+            },
+            {
+                name: 'Teclado',
+            },
+            {
+                name: 'Mouse',
+            },
+            {
+                name: 'Mouse Pads',
+            },
+            {
+                name: 'Micrófonos',
+            },
+            {
+                name: 'Parlantes',
+            },
+            {
+                name: 'Joysticks',
+            },
+            {
+                name: 'Webcams',
+            },
+            {
+                name: 'Conectividad',
+            },
+            {
+                name: 'Tabletas Digitalizadoras',
+            },
+            {
+                name: 'Sillas Gamer',
+            },
+        ],
+    },
+    {
+        name: 'Notebooks / Tablets',
+        isMainCategory: true,
+    },
+    {
+        name: 'Ofertas',
+        isMainCategory: true,
+    },
+    ];
+    const [selectedCategory, setSelectedCategory] = useState('');
+    const [selectedSubcategory, setSelectedSubcategory] = useState('');
     
-  })
-   const [error,setError]= useState({})
-
-  const categoryHandler = (event)=>{
-   
-    const category = event.target.value;
-    setSelectedCategory(category);
-
-    // Buscar las subcategorías correspondientes a la categoría seleccionada
-    const selectedCategoryObj = categories.find((cat) => cat.name === category);
-    const subcategories = selectedCategoryObj.subcategories || [];
-    setSelectedSubcategory(subcategories[0]?.name || '');
-
-  }
-
-  const handleSubcategory = (event) => {
-    const subcategory = event.target.value;
-    setSelectedSubcategory(subcategory);
-  };
-  
-
- 
- 
-   
-
-  const onChangeHandler =(event)=>{
     
-  //  console.log(event.target.name)
-  setError(
-    validate({
-      ...form, [event.target.name]: event.target.value
+    const [form, setForm] = useState({
+        name:"",
+        description:"",
+        manufacturer: "",
+        origin: "",
+        price:"",
+        discount:0,
+        stock: 0,
+        isActive:false,//fata input
+        category:"",
+        image:"",
+        
     })
-  )
-  
-   setForm({
-    ...form, [event.target.name]: event.target.value
-   })
-  }
-  const onSubmithandler=(event)=>{
-    event.preventDefault();
-    console.log(form)
-    if(Object.keys(error).length)
-      return alert('missing info')
-      
-   
+    const [error,setError]= useState({})
+
+    const categoryHandler = (event)=>{
     
-    axios.post(URL,form)
-    .then(res=>alert(res.data))
-    .catch(error=>alert(error.data))
-    
+        const category = event.target.value;
+        setSelectedCategory(category);
+
+        // Buscar las subcategorías correspondientes a la categoría seleccionada
+        const selectedCategoryObj = categories.find((cat) => cat.name === category);
+        const subcategories = selectedCategoryObj.subcategories || [];
+        setSelectedSubcategory(subcategories[0]?.name || '');
+
     }
+
+    const handleSubcategory = (event) => {
+        const subcategory = event.target.value;
+        setSelectedSubcategory(subcategory);
+    };
+    
+
+    
+    
+    
+
+    const onChangeHandler =(event)=>{
+        
+    //  console.log(event.target.name)
+    setError(
+        validate({
+        ...form, [event.target.name]: event.target.value
+        })
+    )
+    
+    setForm({
+        ...form, [event.target.name]: event.target.value
+    })
+    }
+    const onSubmithandler=(event)=>{
+        event.preventDefault();
+        console.log(form)
+        if(Object.keys(error).length)
+        return alert('missing info')
+        
+    
+        
+        axios.post(URL,form)
+        .then(res=>alert(res.data))
+        .catch(error=>alert(error.data))
+        
+        }
     
   
 
     return (
-    <form action="/products" method="POST" enctype="multipart/form-data" onSubmit={onSubmithandler}>
+    <form action="/products" method="POST" encType="multipart/form-data" onSubmit={onSubmithandler}>
         <div>
             <div>
                 <div>
